Remove previous impl
