+++ gui/package.json
@@ -40,7 +40,6 @@
     "react-router": "^7.12.0",
     "react-use-measure": "^2.1.1",
     "stan-language-server": "^0.4.6",
-    "stanc3": "2.38.0",
     "tinystan": "^0.3.2",
     "webr": "^0.5.9"
   },
