+++ gui/src/app/areas/ModelDataArea/ModelEditorPanel.tsx
@@ -1,25 +1,12 @@
-import { AutoFixHigh, Cancel, Help, Settings } from "@mui/icons-material";
-
-import { Split } from "@geoffcox/react-splitter";
-import { useMonaco } from "@monaco-editor/react";
+import { AutoFixHigh, Help, Settings } from "@mui/icons-material";
 
 import TextEditor from "@SpComponents/FileEditor/TextEditor";
 import { ToolbarItem } from "@SpComponents/FileEditor/ToolBar";
-import useStanc from "@SpCore/Stanc/useStanc";
 import { CompileContext } from "@SpCore/Compilation/CompileContextProvider";
-import {
-  FunctionComponent,
-  use,
-  useCallback,
-  useEffect,
-  useMemo,
-  useState,
-} from "react";
+import { FunctionComponent, use, useCallback, useMemo } from "react";
 import { ProjectContext } from "@SpCore/Project/ProjectContextProvider";
 import { ProjectKnownFiles } from "@SpCore/Project/ProjectDataModel";
 import { FileNames } from "@SpCore/Project/FileMapping";
-
-import StanCompileResultPanel from "./StanCompileResultPanel";
 
 const ModelEditorPanel: FunctionComponent = () => {
   const { data, update } = use(ProjectContext);
@@ -42,50 +29,8 @@ const ModelEditorPanel: FunctionComponent = () => {
     [update],
   );
 
-  const { stancErrors } = useStanc(
-    "main.stan",
-    data.ephemera.stanFileContent,
-    setEditedFileContent,
-  );
-
-  const {
-    compileStatus,
-    compileMessage,
-    compile,
-    validSyntax,
-    setValidSyntax,
-    isConnected,
-  } = use(CompileContext);
-
-  // set syntax validitiy based on what the LSP returned to the editor
-  const monacoInstance = useMonaco();
-  const checkMarkers = useCallback(() => {
-    if (!monacoInstance) return;
-    const error = monacoInstance.editor
-      .getModelMarkers({
-        resource: monacoInstance.Uri.file(FileNames.STANFILE),
-      })
-      .some(
-        ({ source, severity }) =>
-          source === "stan-language-server" &&
-          severity === monacoInstance.MarkerSeverity.Error,
-      );
-    setValidSyntax(!error);
-  }, [monacoInstance, setValidSyntax]);
-
-  useEffect(() => {
-    if (!monacoInstance) return;
-    const disposable = monacoInstance.editor.onDidChangeMarkers(checkMarkers);
-    return () => {
-      disposable.dispose();
-    };
-  }, [checkMarkers, monacoInstance]);
-
-  const hasWarnings = useMemo(() => {
-    return stancErrors.warnings && stancErrors.warnings.length > 0;
-  }, [stancErrors]);
-
-  const [syntaxWindowVisible, setSyntaxWindowVisible] = useState(false);
+  const { compileStatus, compileMessage, compile, validSyntax, isConnected } =
+    use(CompileContext);
 
   const toolbarItems: ToolbarItem[] = useMemo(() => {
     const ret: ToolbarItem[] = [];
@@ -101,28 +46,11 @@ const ModelEditorPanel: FunctionComponent = () => {
     // invalid syntax
     if (!validSyntax && !!data.ephemera.stanFileContent) {
       ret.push({
-        type: "button",
-        icon: <Cancel />,
+        type: "text",
         label: "Syntax error",
         color: "error.dark",
-        tooltip: "Syntax error in Stan file",
-        onClick: () => {
-          setSyntaxWindowVisible((v) => !v);
-        },
-      });
-    } else if (hasWarnings && !!data.ephemera.stanFileContent) {
-      ret.push({
-        type: "button",
-        icon: <Cancel />,
-        label: "Compiler warning",
-        color: "info.dark",
-        tooltip: "Warning in Stan file",
-        onClick: () => {
-          setSyntaxWindowVisible((v) => !v);
-        },
       });
     }
-
     // auto format
     if (data.ephemera.stanFileContent && validSyntax) {
       ret.push({
@@ -168,7 +96,6 @@ const ModelEditorPanel: FunctionComponent = () => {
     validSyntax,
     data.ephemera.stanFileContent,
     data.stanFileContent,
-    hasWarnings,
     compileStatus,
     isConnected,
     compile,
@@ -177,16 +104,7 @@ const ModelEditorPanel: FunctionComponent = () => {
 
   const isCompiling = compileStatus === "compiling";
 
-  const compileResultWindow = syntaxWindowVisible ? (
-    <StanCompileResultPanel
-      stancErrors={stancErrors}
-      onClose={() => setSyntaxWindowVisible(false)}
-    />
-  ) : (
-    <></>
-  );
-
-  const editor = (
+  return (
     <TextEditor
       language="stan"
       label={FileNames.STANFILE}
@@ -198,17 +116,6 @@ const ModelEditorPanel: FunctionComponent = () => {
       toolbarItems={toolbarItems}
       contentOnEmpty="Begin editing or select an example from the left panel"
     />
-  );
-
-  return (
-    <Split
-      horizontal
-      initialPrimarySize={syntaxWindowVisible ? "60%" : "100%"}
-      splitterSize={syntaxWindowVisible ? "7px" : "0px"}
-    >
-      {editor}
-      {compileResultWindow}
-    </Split>
   );
 };
